requirements.txt
cloudpickle==3.0.0
dill==0.3.8
streamlit==1.40.1
langchain==0.3.12
huggingface_hub==0.24.6
faiss-cpu==1.9.0
jinja2==3.1.4 
joblib==1.4.2
langchain-community==0.3.12 
langchain-core==0.3.25
langsmith==0.1.147
mpmath==1.3.0 
numpy==1.26.4
pandas==2.2.3 
pydantic==2.8.2
python-dotenv==0.21.0
requests==2.32.3 
sentence-transformers==3.3.1
scikit-learn==1.5.1
scipy==1.14.1
sys
torch==2.5.1
transformers==4.45.2
